Finalizado projeto

+++ src/main/java/com/example/demo/repository/CarrinhoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Carrinho;
 import com.example.demo.models.Cliente;
 
-public interface CarrinhoRepository extends JpaRepository<Carrinho, Long>{
+public interface CarrinhoRepository extends JpaRepository<Carrinho, Integer>{
 
 }
